+++ lib/objects/recap.g.dart
@@ -7,19 +7,19 @@ part of 'recap.dart';
 // **************************************************************************
 
 Recap _$RecapFromJson(Map<String, dynamic> json) => Recap(
-      poids0: (json['poids0'] as num).toDouble(),
-      date0: DateTime.parse(json['date0'] as String),
-      poids1: (json['poids1'] as num).toDouble(),
-      date1: DateTime.parse(json['date1'] as String),
+      poidsRecent: (json['poidsRecent'] as num).toDouble(),
+      dateRecent: DateTime.parse(json['dateRecent'] as String),
+      poidsLate: (json['poidsLate'] as num).toDouble(),
+      dateLate: DateTime.parse(json['dateLate'] as String),
       difference: (json['difference'] as num).toDouble(),
       etat: $enumDecode(_$EtatPoidsEnumMap, json['etat']),
     );
 
 Map<String, dynamic> _$RecapToJson(Recap instance) => <String, dynamic>{
-      'poids0': instance.poids0,
-      'date0': instance.date0.toIso8601String(),
-      'poids1': instance.poids1,
-      'date1': instance.date1.toIso8601String(),
+      'poidsRecent': instance.poidsRecent,
+      'dateRecent': instance.dateRecent.toIso8601String(),
+      'poidsLate': instance.poidsLate,
+      'dateLate': instance.dateLate.toIso8601String(),
       'difference': instance.difference,
       'etat': _$EtatPoidsEnumMap[instance.etat],
     };